fix(bug fix): modify flow of functions in asyncronous way
Fix bug that console does not work as intended

Problem:
- MissionUtils.Console.readLine is based on readline api
- Readline function works on asyncronous way
- As the code was written in syncronous way, it didn't as intended

Solution:
- Collocate all the functions inside of the callback of readLine method
- Use recursion of readLine method instead of using while loop

Modifications:
- Delete Game class
- Put the logic of getting answer from user and matching with the question inside of the callback of getUserAnswer method of Input class
- Make getUserAnswer method of Input class call itself when the game needs next input from user
- Make getReplayRequest mehtod of Input class call getUserAnswer method with new question when user wants to play a new game
- Remove all readline closing from Input and Output class
- Add MissionUtils.Console.close at quitting game part which is the end of getReplayRequest method
