+++ src/App.js
@@ -1,4 +1,4 @@
-const Game = require("../src/Game");
+const Question = require("../src/Question");
 const Input = require("../src/Input");
 const { Output, WELCOME_MESSAGE } = require("../src/Output");
 
@@ -6,13 +6,7 @@ class App {
   play() {
     Output.printToUser(WELCOME_MESSAGE);
 
-    while (true) {
-      const game = new Game();
-      game.run();
-
-      const userWantsReplay = Input.getReplayRequest();
-      if (!userWantsReplay) break;
-    }
+    Input.getUserAnswer(Question.create());
   }
 }
