+++ src/Output.js
@@ -6,7 +6,6 @@ const END_MESSAGE = "3개의 숫자를 모두 맞히셨습니다! 게임 종료"
 class Output {
   static printToUser(message) {
     MissionUtils.Console.print(message);
-    MissionUtils.Console.close();
   }
 }
